# docs/swagger.yaml
definitions:
...
  dto.CreateProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        type: array
      is_posted:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      total:
        type: integer
    type: object
  dto.Product:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_boosted:
        type: boolean
      is_posted:
        type: boolean
      name:
        type: string
      price:
        type: number
      seller_id:
        type: integer
      stock:
        type: integer
      store_id:
        type: integer
    type: object
...
  dto.UpdateProductRequest:
    properties:
      category:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
...
  errorx.AppError:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        $ref: '#/definitions/errorx.ErrorCode'
      message:
        type: string
    type: object
  errorx.ErrorCode:
    enum:
    - UNAUTHORIZED
    - FORBIDDEN
    - VALIDATION_ERROR
    - NOT_FOUND
    - INTERNAL_ERROR
    - BAD_REQUEST
    - CONFLICT
    type: string
    x-enum-varnames:
    - ErrUnauthorized
    - ErrForbidden
    - ErrValidation
    - ErrNotFound
    - ErrInternal
    - ErrBadRequest
    - ErrConflict
  response.BaseResponse:
    properties:
      data: {}
      error: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /products:
    get:
      description: Fetch all available products across stores with filtering/pagination
      parameters:
      - description: Category
        in: query
        name: category
        type: string
      - description: Search query
        in: query
        name: query
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: List all products (public)
      tags:
      - Product
  /products/{id}:
    get:
      description: Fetch details of a single product (public endpoint)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Get a product by ID (Public)
      tags:
      - Product
...
  /store/:store_id/product:
    post:
      consumes:
      - application/json
      description: Add a new product to a store (Admin only)
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      - description: Product details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Add product
      tags:
      - Product
  /store/:store_id/product/:id:
    delete:
      description: Completely remove a product from a store (Admin only)
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Delete product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Modify details of an existing product (Admin only)
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Edit product
      tags:
      - Product
  /store/:store_id/products:
    get:
      description: Retrieve a paginated list of all products for a store
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: List products
      tags:
      - Product
...
swagger: "2.0"